@@ -1,6 +1,6 @@
 package com.wallpaper.tim.phoneinsidewallpaper;
 
-import android.support.v4.app.FragmentActivity;
+import androidx.fragment.app.FragmentActivity;
 import android.os.Bundle;
 
 public class MainActivity extends FragmentActivity {
