@@ -1,7 +1,7 @@
 package com.wallpaper.tim.phoneinsidewallpaper.UI;
 
 import android.os.Bundle;
-import android.support.v4.app.FragmentActivity;
+import androidx.fragment.app.FragmentActivity;
 import android.view.View;
 
 import com.wallpaper.tim.phoneinsidewallpaper.R;
